+++ spec/AssetSpec.js
@@ -68,38 +68,4 @@ describe("test Asset", function() {
 		expect(asset.destroyed()).toBe(true);
 		expect(asset._system).toBeUndefined();
 	});
-
-	it("asSurface", function(done) {
-		var runtime = skeletonRuntime();
-		var scene = runtime.scene;
-		var surface = new mock.Surface(1, 1);
-		expect(g.ImageAsset.asSurface(surface)).toBe(surface);
-
-		var undefinedScene = false;
-		expect(g.ImageAsset.asSurface(undefinedScene)).toBe(false);
-
-		var game = new mock.Game({
-			width: 320,
-			height: 270,
-			assets: {
-				foo: {
-					type: "image",
-					path: "/dummypath.png",
-					virtualPath: "dummypath.png",
-					global: true,
-					width: 1,
-					height: 1
-				}
-			}
-		});
-
-		game._loaded.add(function () {
-			expect(g.ImageAsset.asSurface(game.assets.foo)).toEqual(game.assets.foo.asSurface());
-			done();
-		});
-		game._startLoadingGlobalAssets();
-
-		jasmine.addMatchers(require("./helpers/customMatchers"));
-		expect(function () { g.ImageAsset.asSurface(scene); }).toThrowError("TypeMismatchError");
-	});
 });
