+++ src/Pane.ts
@@ -7,6 +7,7 @@ import { Renderer } from "./Renderer";
 import { EntityStateFlags } from "./EntityStateFlags";
 import { Camera } from "./Camera";
 import { Matrix } from "./Matrix";
+import { SurfaceUtil } from "./SurfaceUtil";
 
 /**
  * `Pane` のコンストラクタに渡すことができるパラメータ。
@@ -123,7 +124,7 @@ export class Pane extends CacheableE {
 		super(param);
 		this._oldWidth = param.width;
 		this._oldHeight = param.height;
-		this.backgroundImage = ImageAsset.asSurface(param.backgroundImage);
+		this.backgroundImage = SurfaceUtil.asSurface(param.backgroundImage);
 		this.backgroundEffector = param.backgroundEffector;
 		this._shouldRenderChildren = false;
 		this._padding = param.padding;
